Moved several classes to `core`: - `JobKey` - `ExecutionIndex` - `ExecutionStatus` - `ExecutionStore` - `OutputStore`
