--- backend/src/main/scala/cromwell/backend/BackendExecutionStatus.scala
@@ -1,6 +1,0 @@
-package cromwell.backend
-
-object BackendExecutionStatus extends Enumeration {
-  type BackendExecutionStatus = Value
-  val NotStarted, Starting, Running, Failed, Preempted, Done, Aborted = Value
-}
