+++ backend/src/main/scala/cromwell/backend/BackendLifecycleActorFactory.scala
@@ -5,6 +5,7 @@ import java.nio.file.Path
 import akka.actor.Props
 import com.typesafe.config.Config
 import cromwell.backend.io.WorkflowPaths
+import cromwell.core.{ExecutionStore, OutputStore}
 import wdl4s.Call
 import wdl4s.expression.WdlStandardLibraryFunctions
 
@@ -16,8 +17,8 @@ trait BackendLifecycleActorFactory {
 
   def workflowFinalizationActorProps(workflowDescriptor: BackendWorkflowDescriptor,
                                      calls: Seq[Call],
-                                     executionStore: BackendExecutionStore,
-                                     outputStore: BackendOutputStore): Option[Props] = None
+                                     executionStore: ExecutionStore,
+                                     outputStore: OutputStore): Option[Props] = None
 
   def expressionLanguageFunctions(workflowDescriptor: BackendWorkflowDescriptor,
                                   jobKey: BackendJobDescriptorKey): WdlStandardLibraryFunctions
