+++ core/src/main/scala/cromwell/core/ExecutionStore.scala
@@ -1,0 +1,14 @@
+package cromwell.core
+
+import cromwell.core.ExecutionStatus._
+
+import scala.language.postfixOps
+
+object ExecutionStore {
+  def empty = ExecutionStore(Map.empty)
+  type ExecutionStoreEntry = (JobKey, ExecutionStatus)
+}
+
+case class ExecutionStore(store: Map[JobKey, ExecutionStatus]) {
+  def add(values: Map[JobKey, ExecutionStatus]) = this.copy(store = store ++ values)
+}
