+++ core/src/main/scala/cromwell/core/JobKey.scala
@@ -1,0 +1,10 @@
+package cromwell.core
+
+import wdl4s.Scope
+
+trait JobKey {
+  def scope: Scope
+  def index: Option[Int]
+  def attempt: Int
+  def tag: String
+}
