+++ engine/src/main/scala/cromwell/engine/backend/OldStyleJobDescriptor.scala
@@ -1,7 +1,8 @@
 package cromwell.engine.backend
 
-import cromwell.backend.{ExecutionHash, JobKey}
+import cromwell.backend.ExecutionHash
 import cromwell.backend.wdl.OldCallEngineFunctions
+import cromwell.core.JobKey
 import cromwell.engine.AbortRegistrationFunction
 import cromwell.engine.backend.runtimeattributes.CromwellRuntimeAttributes
 import cromwell.engine.workflow.{BackendCallKey, FinalCallKey}
