+++ engine/src/main/scala/cromwell/engine/workflow/lifecycle/CopyWorkflowOutputsActor.scala
@@ -4,10 +4,9 @@ import java.nio.file.Path
 
 import akka.actor.Props
 import cromwell.backend.{BackendConfigurationDescriptor, BackendLifecycleActorFactory}
-import cromwell.core.{PathCopier, WorkflowId}
+import cromwell.core.{ExecutionStore, OutputStore, PathCopier, WorkflowId}
 import cromwell.engine.EngineWorkflowDescriptor
 import cromwell.engine.backend.{BackendConfiguration, CromwellBackends}
-import cromwell.engine.workflow.lifecycle.execution.{ExecutionStore, OutputStore}
 import wdl4s.ReportableSymbol
 import wdl4s.values.WdlSingleFile
