+++ engine/src/main/scala/cromwell/webservice/MetadataBuilderActor.scala
@@ -4,9 +4,9 @@ import java.sql.Timestamp
 import java.time.OffsetDateTime
 
 import akka.actor.{ActorRef, LoggingFSM, Props}
-import cromwell.core.{WorkflowId, WorkflowState}
+import cromwell.core.{ExecutionIndex, WorkflowId, WorkflowState}
 import cromwell.database.obj.WorkflowMetadataKeys
-import cromwell.engine.ExecutionIndex.ExecutionIndex
+import ExecutionIndex.ExecutionIndex
 import cromwell.services.MetadataServiceActor._
 import cromwell.services.ServiceRegistryActor.ServiceRegistryFailure
 import cromwell.services._
